backend/requirements.txt
# Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10

# HTTP and Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
youtube-transcript-api==0.6.1
lxml==4.9.3

# AI/LLM Integration
anthropic==0.3.11

# Environment and Configuration
python-decouple==3.8
python-dotenv==1.0.0

# Database (using existing database_handler)
psycopg2-binary>=2.9.0

# File Processing
PyPDF2==3.0.1

# Validation and Utilities
validators==0.21.2
Werkzeug==2.3.7

# Development and Testing
pytest==7.4.2
pytest-flask==1.2.0

# Logging
structlog>=22.0.0